feat: replace CollectionsControls with CollectionsActions and add CollectionsLayers
