+++ apps/web/i18n/locales/en-US.json
@@ -252,7 +252,7 @@
       "rssUrl": "RSS URL"
     },
 
-    "controls": {
+    "layers": {
       "properties": {
         "title": "Display properties",
         "link": "Link",
